Update blockquote
Fixed weird line on blockquote background; still having trouble with quotation mark positioning (try inserting quotation marks with html instead of css? but that would mean having to add them manually to every blockquote; maybe instead, add another layer of containing divs? will look into further) and with changing the width of the div on mobile (come to think of it, this problem might solve itself when I add the grey-and-white styling; will give that a try first)